fixes to bring flud back to life.
This push has several fixes to bring flud back into working order against recent twisted / pycrypto changes, as well as reformats the code base to use spaces instead of tabs.

There is a fix for HTTPMultipartDownloader, who's interface now expects it to have a 'file' member variable.

There is a fix for random number generation, which requires atfork() to be called if used in a thread context for security.

There is a fix for accessing RSAobj, which is now named _RSAobj.

A '-h' option is added to metadataviewer to get hex output.

This code now runs correctly on the latest ubuntu release as of today, 11.04. Untested on other platforms.